+++ tests/test_live.sh
@@ -29,6 +29,7 @@
 #   cortex unack_error, cortex notifications_recent, cortex file_tails, cortex notifications_test,
 #   cortex llm_invocations,
 #   cortex similar_incidents, cortex ask_history, cortex incident_context, cortex graph,
+#   cortex skill_events,
 #   cortex help
 #
 # Exit codes: